+++ test/lib/absinthe_test.exs
@@ -2,13 +2,28 @@ defmodule AbsintheTest do
   use Absinthe.Case, async: true
   import AssertResult
 
+  it "can return multiple errors" do
+    query = "mutation { failingThing(type: MULTIPLE) { name } }"
+    assert_result {:ok, %{data: %{"failingThing" => nil}, errors: [%{message: "In field \"failingThing\": one"}, %{message: "In field \"failingThing\": two"}]}}, run(query, Things)
+  end
+
   it "can return extra error fields" do
-    query = "mutation { FailingThing(type: WITH_CODE) { name } }"
-    assert_result {:ok, %{data: %{}, errors: [%{code: 42, message: "In field \"FailingThing\": Custom Error"}]}}, run(query, Things)
+    query = "mutation { failingThing(type: WITH_CODE) { name } }"
+    assert_result {:ok, %{data: %{"failingThing" => nil}, errors: [%{code: 42, message: "In field \"failingThing\": Custom Error"}]}}, run(query, Things)
   end
 
   it "requires message in extended errors" do
     query = "mutation { FailingThing(type: WITHOUT_MESSAGE) { name } }"
+    assert_raise Absinthe.ExecutionError, fn -> run(query, Things) end
+  end
+
+  it "can return multiple errors, with extra error fields" do
+    query = "mutation { failingThing(type: MULTIPLE_WITH_CODE) { name } }"
+    assert_result {:ok, %{data: %{"failingThing" => nil}, errors: [%{code: 1, message: "In field \"failingThing\": Custom Error 1"}, %{code: 2, message: "In field \"failingThing\": Custom Error 2"}]}}, run(query, Things)
+  end
+
+  it "requires message in extended errors, when multiple errors are given" do
+    query = "mutation { failingThing(type: MULTIPLE_WITHOUT_MESSAGE) { name } }"
     assert_raise Absinthe.ExecutionError, fn -> run(query, Things) end
   end
 
@@ -59,6 +74,18 @@ defmodule AbsintheTest do
     }
     """
     assert_result {:ok, %{data: %{"things" => [%{"name" => "Bar", "id" => "bar"}, %{"name" => "Foo", "id" => "foo"}]}}}, run(query, Things)
+  end
+
+  it "Invalid arguments on children of a list field are correctly handled" do
+    query = """
+    query AllTheThings {
+      things {
+        id(x: 1)
+        name
+      }
+    }
+    """
+    assert_result {:ok, %{errors: [%{message: "Unknown argument \"x\" on field \"id\" of type \"Thing\"."}]}}, run(query, Things)
   end
 
   it "can do a simple query with an all caps alias" do
@@ -155,7 +182,7 @@ defmodule AbsintheTest do
       }
     """
     assert_result {:ok, %{data: %{"thingByContext" => %{"name" => "Bar"}}}}, run(query, Things, context: %{thing: "bar"})
-    assert_result {:ok, %{data: %{},
+    assert_result {:ok, %{data: %{"thingByContext" => nil},
                           errors: [%{message: ~s(In field "thingByContext": No :id context provided)}]}}, run(query, Things)
   end