+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## v1.2.6
+
+- Enhancement: Query complexity analysis!
+- Bug fix: Fields with runtime errors are presented with `null` values in the result data instead of elided entirely.
+
+## v1.2.5
+
+- Enhancement: Scalar type parse functions can access the context. This enables
+uploaded files with Absinthe.Plug
+
+## v1.2.4
+- Enhancement: Complex errors. You can now return `{:error, %{message: "...", other_key: value}}`
+- Bug Fix: Invalid Arguments on a field that is under a list field doesn't error.
+- Bug Fix: Deeper fragment merging.
+
+## v1.2.3
+- Bug Fix: When there are no arguments, an empty map should be passed to the resolution functions not `nil`
+
 ## v1.2.2
 - Enhancement: Enable `import_fields` for input objects. In the future we will
 enforce that `input_objects` can only import fields from other `input_objects`.